refactor: split linker metadata types
